requirements.txt
Django==1.10.5
django-ckeditor
